+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.2.2] - 2020-01-29
 
 ### Added
 
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## [0.2.1] - 2020-01-16
 
-## Changed
+### Changed
 
 - `NameOf()` and `TypeOf()` now produce a more meaningful panic message when passed a `nil` message
 
@@ -57,6 +57,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.1.1]: https://github.com/dogmatiq/configkit/releases/v0.1.1
 [0.2.0]: https://github.com/dogmatiq/configkit/releases/v0.2.0
 [0.2.1]: https://github.com/dogmatiq/configkit/releases/v0.2.1
+[0.2.2]: https://github.com/dogmatiq/configkit/releases/v0.2.2
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD